% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiskewBART.R
\name{MultiskewBART}
\alias{MultiskewBART}
\title{MCMC options for skewBART}
\usage{
MultiskewBART(
  X = X_norm,
  Y = Y_scaled,
  test_X = test_X_norm,
  tree_num = 200,
  iter = 5000,
  burn = 2500
)
}
\arguments{
\item{X}{NxP matrix of training data covariates.}

\item{Y}{Nxk matrix of training data response.}

\item{test_X}{MxP matrix of test data covariates}

\item{tree_num}{the number of trees}

\item{iter}{Number of MCMC iteration}

\item{burn}{Number of warmup iterations for the chain.}
}
\value{
Returns a list with the following components:
\itemize{
  \item y_hat_train: fit to the training data for each iteration of the chain
  \item y_hat_test: fit to the testing data for each iteration of the chain
  \item Sigma: posterior samples of the covariance matrix
  \item lambda: posterior samples of skewness matrix
}
}
\description{
MCMC options for skewBART
}
